backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.1
sqlmodel==0.0.8
pydantic==1.10.18
pandas==2.2.3
joblib==1.3.2
python-dateutil==2.8.2
pytest==7.4.2
httpx==0.24.1



